fix signed conversion warnings

+++ src/private/std_array.c
@@ -24,7 +24,7 @@ set_capacity(
     memcpy(
         AGC_CAST(void, &a->gc, bi),
         AGC_CAST(void, &a->gc, o->buff.v.heap_idx),
-        o->sz * sizeof(avalue_t));
+        (size_t)o->sz * sizeof(avalue_t));
     o->cap = cap;
     av_collectable(&o->buff, AVT_FIXED_BUFFER, bi);
 }
